split inception example to other pr
